DL-12653 review comments
