+++ it/tests/helpers/WireMockMethods.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package helpers
+package tests.helpers
 
 import com.github.tomakehurst.wiremock.client.MappingBuilder
 import com.github.tomakehurst.wiremock.client.WireMock._
